[Black Hole Tech] Quasar cannon.
Adding a first weapon of the black hole tech series - the Quasar.
3 seconds warmup and cooldown
Shoots beam of matter by generating a black hole in front of itself using it's stabilizers, like a natural quasar in space but much smaller. Advanced recuperation system reduces energy consumption, making the user able to fire it for a longer time.